[finishes #161590731] Create Parcel delivery order feature
Write html for the parcel delivery order , and add a table
to capture the details of the parcel.

Create an accordion that will collapse the columns when the screen
is less than 475px wide to make it look better on smaller screens
and to take up space of only the feature that i being used.

Add css that will make the page resonsive and more mobile friendly.